Reorganized express backend
